requirements.txt
numpy==1.26.0
Pillow==10.0.1
pydicom==2.4.3
pynetdicom==2.0.2
